feat: tratamento de erros para get, put e delete by id endpoint publicaitions

+++ src/publications/publications.service.ts
@@ -1,4 +1,4 @@
-import { Inject, Injectable, NotFoundException, forwardRef } from '@nestjs/common';
+import { ForbiddenException, Inject, Injectable, NotFoundException, forwardRef } from '@nestjs/common';
 import { CreatePublicationDto } from './dto/create-publication.dto';
 import { UpdatePublicationDto } from './dto/update-publication.dto';
 import { PublicationsRepository } from './publications.repository';
@@ -16,12 +16,21 @@ export class PublicationsService {
     private readonly mediasService: MediasService,
   ) { }
 
+  private checkPreviousDate(pubDate: Date): boolean {
+    const todaysDate: Date = new Date();
+    return (pubDate < todaysDate);
+  }
+
+  private async checkPubNeededIds(postId : number, mediaId: number) {
+    await this.postsService.getPostbyId(postId);
+    await this.mediasService.getMediaById(mediaId);
+  }
+
   async createPub(body: CreatePublicationDto) {
     const { mediaId, postId } = body;
     // como puxa as funcoes da service o throw ja eh feito
     // a partir da propria
-    await this.postsService.getPostbyId(postId);
-    await this.mediasService.getMediaById(mediaId);
+    this.checkPubNeededIds(postId, mediaId);
     return await this.pubRepository.createPub(body);
   }
 
@@ -59,10 +68,18 @@ export class PublicationsService {
   }
 
   async updatePubById(id: number, body: UpdatePublicationDto) {
+    const checkPub = await this.pubRepository.getPubById(id);
+    if(checkPub.length === 0) throw new NotFoundException(`there is no publication for given id=${id}`);
+
+    if(this.checkPreviousDate(checkPub[0].date)) throw new ForbiddenException(`publciaiton with given id=${id} has already been published`)
+    this.checkPubNeededIds(checkPub[0].postId, checkPub[0].mediaId);
+
     return await this.pubRepository.updatePubById(id, body);
   }
 
   async deletePubById(id: number) {
+    const checkPub = await this.pubRepository.getPubById(id);
+    if(checkPub.length === 0) throw new NotFoundException(`there is no publication for given id=${id}`);
     return await this.pubRepository.deletePubById(id);
   }
 }
